code/requirements.txt
fastapi[all]==0.116.1
uvicorn==0.35.0
psycopg2-binary
SQLAlchemy[asyncio]==2.0.38
requests==2.32.3
openpyxl==3.1.5
redis==5.2.1
python-dotenv==1.0.1
rsa==4.9
PyJWT==2.10.1
bitrix24-rest==2.0.3
Jinja2==3.1.6
asyncio==3.4.3
pillow==11.2.1
transliterate==1.10.2
vobject==0.9.9
requests==2.32.3
qrcode==8.1
pymongo==4.15.0
elasticsearch==9.1.0
tqdm==4.66.6
ldap3==2.9.1
#locust==2.37.11
aiofiles==24.1.0
fastapi-socketio==0.0.10
asyncpg==0.30.0
